@@ -16,15 +16,18 @@
     "lodash-es": "^4.17.15",
     "pepjs": "^0.5.2",
     "postprocessing": "^6.8.5",
-    "react": "^0.0.0-experimental-5faf377df",
+    "react": "^16.12.0",
     "react-app-polyfill": "^1.0.4",
     "react-dev-utils": "^9.1.0",
-    "react-dom": "^0.0.0-experimental-5faf377df",
+    "react-dom": "^16.12.0",
     "react-promise-suspense": "^0.3.1",
+    "react-reconciler": "^0.24.0",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-spring": "8.0.27",
+    "react-three-gui": "^0.1.5",
     "react-use-gesture": "6.0.12",
+    "scheduler": "^0.18.0",
     "styled-components": "^4.4.0",
     "three": "^0.109.0",
     "three.meshline": "^1.2.0",
