@@ -61,62 +63,71 @@
   },
   "homepage": "https://github.com/drcmda/react-three-fiber#readme",
   "dependencies": {
-    "@babel/runtime": "^7.7.1",
-    "@juggle/resize-observer": "^2.4.0",
-    "react-promise-suspense": "^0.3.1",
-    "react-reconciler": "0.23.0",
+    "@babel/runtime": "^7.7.7",
+    "@juggle/resize-observer": "^3.0.2",
+    "react-merge-refs": "^1.0.0",
+    "react-promise-suspense": "^0.3.2",
+    "react-reconciler": "0.24.0",
     "react-use-measure": "^2.0.0",
     "resize-observer-polyfill": "^1.5.1",
-    "scheduler": "0.17.0",
-    "tiny-emitter": "^2.1.0"
+    "scheduler": "0.18.0",
+    "tiny-emitter": "^2.1.0",
+    "utility-types": "^3.10.0"
   },
   "peerDependencies": {
-    "react": ">=16.11",
-    "react-dom": ">=16.11",
-    "three": ">=0.110"
+    "react": ">=16.12",
+    "react-dom": ">=16.12",
+    "three": ">=0.112"
   },
   "devDependencies": {
-    "@babel/core": "7.7.0",
-    "@babel/plugin-transform-modules-commonjs": "7.7.0",
-    "@babel/plugin-transform-parameters": "7.4.4",
-    "@babel/plugin-transform-runtime": "7.6.2",
-    "@babel/plugin-transform-template-literals": "7.4.4",
-    "@babel/preset-env": "7.7.1",
-    "@babel/preset-react": "7.7.0",
-    "@babel/preset-typescript": "^7.7.0",
+    "@ampproject/rollup-plugin-closure-compiler": "^0.18.1",
+    "@babel/core": "7.7.7",
+    "@babel/plugin-transform-modules-commonjs": "7.7.5",
+    "@babel/plugin-transform-parameters": "7.7.7",
+    "@babel/plugin-transform-runtime": "7.7.6",
+    "@babel/plugin-transform-template-literals": "7.7.4",
+    "@babel/preset-env": "7.7.7",
+    "@babel/preset-react": "7.7.4",
+    "@babel/preset-typescript": "^7.7.7",
+    "@types/jest": "^24.0.25",
     "@types/lodash-es": "^4.17.3",
-    "@types/react": "^16.9.11",
-    "@types/react-native": "^0.60.22",
+    "@types/react": "^16.9.17",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-native": "^0.60.28",
     "@types/react-reconciler": "^0.18.0",
     "@types/scheduler": "^0.16.1",
-    "@typescript-eslint/eslint-plugin": "^2.6.1",
-    "@typescript-eslint/parser": "^2.6.1",
+    "@typescript-eslint/eslint-plugin": "^2.15.0",
+    "@typescript-eslint/parser": "^2.15.0",
     "babel-eslint": "^10.0.3",
+    "@types/react-test-renderer": "^16.9.1",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "copyfiles": "^2.1.1",
-    "cypress": "^3.6.0",
-    "eslint": "^6.6.0",
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^2.2.0",
-    "expo-gl": "^7.0.0",
-    "expo-three": "^5.2.0",
-    "husky": "^3.0.9",
+    "cypress": "^3.8.1",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.9.0",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
+    "expo-gl": "^8.0.0",
+    "expo-three": "^5.3.0",
+    "husky": "^4.0.1",
     "json": "^9.0.6",
-    "lint-staged": "^9.4.2",
+    "lint-staged": "^9.5.0",
     "prettier": "^1.19.1",
-    "react": "^16.11.0",
+    "jest": "^24.9.0",
     "react-dom": "^16.11.0",
+    "react-test-renderer": "^16.12.0",
+    "react": "^16.11.0",
     "rimraf": "^3.0.0",
-    "rollup": "^1.26.5",
+    "rollup": "^1.28.0",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-size-snapshot": "^0.10.0",
     "three": "^0.109.0",
-    "typescript": "^3.7.2",
+    "typescript": "^3.7.4",
+    "ts-jest": "^24.3.0",
     "wait-on": "^3.3.0"
   },
   "collective": {
