@@ -15,5 +15,5 @@ function App() {
   )
 }
 
-ReactDOM.createRoot(document.getElementById('root')).render(<App />)
-//ReactDOM.render(<App />, document.getElementById('root'))
+//ReactDOM.createRoot(document.getElementById('root')).render(<App />)
+ReactDOM.render(<App />, document.getElementById('root'))
