@@ -56,16 +56,17 @@ module.exports = mode => {
       modules: [path.resolve('./'), 'node_modules'],
       extensions: ['.js', '.jsx', '.ts', '.tsx'],
       alias: {
-        'react-three-fiber': path.resolve(`../src/targets/web/index.ts`),
+        'react-three-fiber': path.resolve(`../src/targets/web`),
         //'react-three-fiber/extras': path.resolve(`../src/extras/index.ts`),
-        //scheduler: path.resolve('node_modules/scheduler'),
         react: path.resolve('node_modules/react'),
         'react-dom': path.resolve('node_modules/react-dom'),
+        scheduler: path.resolve('node_modules/scheduler'),
+        'react-scheduler': path.resolve('node_modules/react-scheduler'),
         'prop-types': path.resolve('node_modules/prop-types'),
         //three: path.resolve('node_modules/three'),
         //three$: path.resolve('node_modules/three/src/Three'),
-        three$: path.resolve('./resources/three.js'),
-        '../../../build/three.module.js': path.resolve('./resources/three.js'),
+        //three$: path.resolve('./resources/three.js'),
+        //'../../../build/three.module.js': path.resolve('./resources/three.js'),
         'react-spring/three': createAlias('react-spring/src/targets/three', 'react-spring/three'),
         'react-use-gesture': createAlias('react-use-gesture/index.js', 'react-use-gesture'),
         //'react-use-measure': createAlias('react-use-measure/src/web', '../node_modules/react-use-measure'),
