@@ -2,6 +2,7 @@ import path from 'path'
 import babel from 'rollup-plugin-babel'
 import resolve from 'rollup-plugin-node-resolve'
 import json from 'rollup-plugin-json'
+import compiler from '@ampproject/rollup-plugin-closure-compiler'
 import { sizeSnapshot } from 'rollup-plugin-size-snapshot'
 
 const root = process.platform === 'win32' ? path.resolve('/') : '/'
@@ -35,6 +36,7 @@ function createConfig(entry, out) {
         babel(getBabelOptions({ useESModules: true }, '>1%, not dead, not ie 11, not op_mini all')),
         sizeSnapshot(),
         resolve({ extensions }),
+        //compiler(),
       ],
     },
     {
@@ -47,7 +49,11 @@ function createConfig(entry, out) {
 }
 
 export default [
-  ...createConfig('targets/web/index', 'web'),
+  ...createConfig('targets/web', 'web'),
+  ...createConfig('targets/svg', 'svg'),
+  ...createConfig('targets/css2d', 'css2d'),
+  ...createConfig('targets/css3d', 'css3d'),
   //...createConfig('extras/index', 'extras'),
   ...createConfig('targets/native/index', 'native'),
+  ...createConfig('components', 'components'),
 ]
