@@ -17,7 +17,9 @@ export type DomEvent =
   | React.WheelEvent<HTMLDivElement>
   | React.PointerEvent<HTMLDivElement>
 
-export type Intersection = THREE.Intersection & { eventObject: THREE.Object3D }
+export interface Intersection extends THREE.Intersection {
+  eventObject: THREE.Object3D
+}
 
 export type PointerEvent = DomEvent &
   Intersection & {
@@ -61,6 +63,7 @@ export type CanvasContext = SharedCanvasContext & {
   active: boolean
   manual: number
   vr: boolean
+  concurrent: boolean
   invalidateFrameloop: boolean
   frames: number
   subscribers: Subscription[]
@@ -76,10 +79,11 @@ export type ResizeOptions = {
   scroll?: boolean
 }
 
-export type CanvasProps = {
+export interface CanvasProps {
   children: React.ReactNode
   vr?: boolean
   gl2?: boolean
+  concurrent?: boolean
   shadowMap?: boolean | Partial<THREE.WebGLShadowMap>
   orthographic?: boolean
   resize?: ResizeOptions
@@ -98,7 +102,7 @@ export type CanvasProps = {
   onPointerMissed?: () => void
 }
 
-export type UseCanvasProps = CanvasProps & {
+export interface UseCanvasProps extends CanvasProps {
   gl: THREE.WebGLRenderer
   size: RectReadOnly
 }
@@ -130,6 +134,7 @@ export const useCanvas = (props: UseCanvasProps): PointerEvents => {
     size,
     pixelRatio,
     vr = false,
+    concurrent = false,
     shadowMap = false,
     invalidateFrameloop = false,
     updateDefaultCamera = true,
@@ -175,6 +180,7 @@ export const useCanvas = (props: UseCanvasProps): PointerEvents => {
     active: true,
     manual: 0,
     vr,
+    concurrent,
     noEvents,
     invalidateFrameloop: false,
     frames: 0,
@@ -223,8 +229,9 @@ export const useCanvas = (props: UseCanvasProps): PointerEvents => {
     state.current.invalidateFrameloop = invalidateFrameloop
     state.current.vr = vr
     state.current.gl = gl
+    state.current.concurrent = concurrent
     state.current.noEvents = noEvents
-  }, [invalidateFrameloop, vr, noEvents, ready, size, defaultCam, gl])
+  }, [invalidateFrameloop, vr, concurrent, noEvents, ready, size, defaultCam, gl])
 
   // Adjusts default camera
   useMemo(() => {
@@ -486,6 +493,8 @@ export const useCanvas = (props: UseCanvasProps): PointerEvents => {
   // we know we are ready to compile shaders, call subscribers, etc
   const IsReady = useCallback(() => {
     const activate = () => setReady(true)
+
+    // eslint-disable-next-line react-hooks/rules-of-hooks
     useEffect(() => {
       const result = onCreated && onCreated(state.current)
       return void (result && result.then ? result.then(activate) : activate())
@@ -521,8 +530,8 @@ export const useCanvas = (props: UseCanvasProps): PointerEvents => {
   useEffect(
     () => () => {
       if (state.current.gl) {
-        state.current.gl.forceContextLoss!()
-        state.current.gl.dispose!()
+        if (state.current.gl.forceContextLoss) state.current.gl.forceContextLoss!()
+        if (state.current.gl.dispose) state.current.gl.dispose!()
         ;(state.current as any).gl = undefined
         unmountComponentAtNode(state.current.scene)
         state.current.active = false
